CRUD de Produtos com SweetAlert2

+++ database/migrations/2022_08_23_144722_create_produtos_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('nome', 100);
             $table->text('descricao')->nullable();
-            $table->integer('peso')->nullable();
+            $table->double('peso',4,2)->nullable();
             $table->float('preco_venda', 8, 2)->default('0.01');
             $table->integer('estoque_minimo')->default(1);
             $table->integer('estoque_maximo')->default(1);
